yolov5/CNN_Model/requirements.txt
# torch>=1.11.0
# torchvision>=0.12.0
opencv-python
pandas
matplotlib
Pillow
tqdm